Handle multipart/form-data file=@example.wav

+++ src/main/java/Asr.java
@@ -1,12 +1,15 @@
+import com.ibm.websphere.jaxrs20.multipart.IAttachment;
+import com.ibm.websphere.jaxrs20.multipart.IMultipartBody;
 import cz.cuni.mff.ufal.butasr.Client;
 
 import javax.inject.Inject;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
-import java.util.concurrent.CompletableFuture;
+import java.util.Objects;
 import java.util.concurrent.ExecutionException;
 
 @Path("asr")
@@ -17,17 +20,36 @@ public class Asr {
 
     @Path("{lang}")
     @POST
+    @Consumes(MediaType.MULTIPART_FORM_DATA)
     @Produces(MediaType.TEXT_PLAIN)
-    public String getTranscript(@PathParam("lang") String lang, InputStream inputStream){
+    public String getTranscript(@PathParam("lang") String lang, IMultipartBody body){
+        if(body == null){
+            throw new WebApplicationException("No multipart body", Response.Status.BAD_REQUEST);
+        }
         try {
             //System.out.println(file.toAbsolutePath());
-            final java.nio.file.Path file = Files.createTempFile("asr_lid_fileupload_", null);
-            Files.copy(inputStream, file, StandardCopyOption.REPLACE_EXISTING);
-            String transcriptOfWav = client.getTranscriptOfWav(file).get();
-            file.toFile().delete();
-            return transcriptOfWav;
+            final IAttachment attachment = body.getAttachment("file");
+            if(attachment == null){
+                throw new WebApplicationException("The form field with file should be named 'file'," +
+                        " eg. Content-Disposition: form-data; name=\"file\"; filename=\"example.wav\"",
+                        Response.Status.BAD_REQUEST);
+            }
+            try (InputStream inputStream = attachment.getDataHandler().getInputStream()) {
+                java.nio.file.Path file = null;
+                try {
+                    file = Files.createTempFile("asr_lid_fileupload_", null);
+                    // AudioSystem has issues recognizing the wav when passed as InputStream, saving to a tmp file
+                    // There'll probably be some preprocessing in the future
+                    Files.copy(inputStream, file, StandardCopyOption.REPLACE_EXISTING);
+                    return client.getTranscriptOfWav(file).get();
+                } finally {
+                    if(file != null) {
+                        file.toFile().delete();
+                    }
+                }
+            }
         } catch (IOException | InterruptedException | ExecutionException e) {
-            return e.getMessage();
+            throw new WebApplicationException(e, Response.Status.INTERNAL_SERVER_ERROR);
         }
     }
